// src/utils/web3/types.ts
export type JB721TierParams = {
	id?: BigNumber; // undefined for outgoing tier (in launch or adjustTiers tx)
	remainingQuantity?: BigNumber; // undefined for outgoing tier (in launch or adjustTiers tx)

	contributionFloor: BigNumber; // uint128
	lockedUntil: BigNumber;
	initialQuantity: BigNumber; // uint64
	votingUnits: number;
	reservedRate: number;
	reservedTokenBeneficiary: string;
	encodedIPFSUri: string; // encoded link to the rewardTier on IPFS
	allowManualMint: boolean;
	shouldUseBeneficiaryAsDefault: boolean;
	transfersPausable: boolean;
};

export interface JB721PricingParams {
	tiers: JB721TierParams[];
	currency: CurrencyOption;
	decimals: number;
	prices: string;
}

export interface JBTiered721Flags {
	lockReservedTokenChanges: boolean;
	lockVotingUnitChanges: boolean;
	lockManualMintingChanges: boolean;
}

export enum JB721GovernanceType {
	NONE,
	TIERED,
	GLOBAL,
}

export interface JBDeployTiered721DelegateData {
	directory: string;
	name: string;
	symbol: string;
	fundingCycleStore: string;
	baseUri: string;
	tokenUriResolver: string;
	contractUri: string;
	owner: string;
	pricing: JB721PricingParams;
	reservedTokenBeneficiary: string;
	store: string; // IJBTiered721DelegateStore
	flags: JBTiered721Flags;
	governanceType: JB721GovernanceType;
}